backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
mysqlclient==2.2.4
python-dotenv==1.0.1
pydantic==2.6.3
requests==2.31.0
python-multipart==0.0.9
scikit-learn==1.4.1.post1
numpy==1.26.4
pillow==10.2.0
pyzbar==0.1.9
python-magic-bin==0.4.14